+++ js/rollup.config.js
@@ -1,0 +1,73 @@
+import commonjs from '@rollup/plugin-commonjs';
+import resolve from '@rollup/plugin-node-resolve';
+import livereload from 'rollup-plugin-livereload';
+import { terser } from 'rollup-plugin-terser';
+import fs from 'fs';
+import {extname, join} from "path";
+
+const production = !process.env.ROLLUP_WATCH;
+
+const libsDir = join(process.cwd(), 'src', 'libs');
+
+const libs = fs.readdirSync(libsDir).filter(p => extname(p) === ".js")
+
+function serve() {
+	let server;
+
+	function toExit() {
+		if (server) server.kill(0);
+	}
+
+	return {
+		writeBundle() {
+			if (server) return;
+			server = require('child_process').spawn('npm', ['run', 'start', '--', '--dev'], {
+				stdio: ['ignore', 'inherit', 'inherit'],
+				shell: true
+			});
+
+			process.on('SIGTERM', toExit);
+			process.on('exit', toExit);
+		}
+	};
+}
+
+export default [
+	...libs.map(libFile=>({
+		input: `${libsDir}/${libFile}`,
+		output: {
+			sourcemap: true,
+			format: 'iife',
+			file: `../priv/${libFile}`,
+			name: 'window',
+			extend: true,
+		},
+		plugins: [
+
+			// If you have external dependencies installed from
+			// npm, you'll most likely need these plugins. In
+			// some cases you'll need additional configuration -
+			// consult the documentation for details:
+			// https://github.com/rollup/plugins/tree/master/packages/commonjs
+			resolve({
+				browser: true
+			}),
+			commonjs(),
+
+			// In dev mode, call `npm run start` once
+			// the bundle has been generated
+			!production && serve(),
+
+			// Watch the `public` directory and refresh the
+			// browser on changes when not in production
+			!production && livereload('public'),
+
+			// If we're building for production (npm run build
+			// instead of npm run dev), minify
+			production && terser()
+		],
+		watch: {
+			clearScreen: false
+		}
+	}))
+];
