+++ priv/bert.js
@@ -1,171 +1,2 @@
-
-// API
-
-function tuple() { return { t: 104, v: Array.apply(null, arguments) }; }
-function list() { return { t: 108, v: Array.apply(null, arguments) }; }
-function map() { return { t: 116, v: Array.apply(null, arguments) }; }
-function atom(o) { return { t: 100, v: utf8_enc(o) }; }
-function string(o) { return { t: 107, v: utf8_enc(o) }; }
-function float(o) { return { t: 70, v: o }; }
-function number(o) {
-  var s, isInteger = (o % 1 === 0);
-  if (isInteger && o >= 0 && o < 256) { return { t: 97, v: o };  }
-  if (isInteger && o >= -134217728 && o <= 134217727) { return {t: 98, v: o}; }
-  return {t: 110, v: o}; }
-function bin(o) {
-  return { t: 109, v: o instanceof ArrayBuffer ? new Uint8Array(o) :
-                      o instanceof Uint8Array  ? o : utf8_enc(o) }; }
-
-// ENCODER
-
-function enc(o) { return fl([131, ein(o)]); }
-function ein(o) { return Array.isArray(o) ? en_108({ t: 108, v: o }) : eval('en_' + o.t)(o); }
-function en_undefined(o) { return [106]; }
-function en_70(o) {
-  var x = Array(8).fill(0).flat();
-  write_Float(x,o.v,0,false,52,8);
-  return [70].concat(x);
-}
-function en_97(o) { return [97, o.v]; }
-function en_98(o) { return [98, o.v >>> 24, (o.v >>> 16) & 255, (o.v >>> 8) & 255, o.v & 255]; }
-function en_99(o) {
-  var obj = o.v.toExponential(20),
-      match = /([^e]+)(e[+-])(\d+)/.exec(obj),
-      exponentialPart = match[3].length == 1 ? "0" + match[3] : match[3],
-      num = Array.from(bin(match[1] + match[2] + exponentialPart).v);
-  return [o.t].concat(num).concat(Array(31 - num.length).fill(0).flat());
-}
-function en_106(o) { return [106]; }
-function en_115(o) { return [115, o.v.length, ar(o)]; }
-function en_119(o) { return [119, ar(o).length, ar(o)]; }
-function en_118(o) { return [118, ar(o).length >>> 8, ar(o).length & 255, ar(o)]; }
-function en_100(o) { return [100, o.v.length >>> 8, o.v.length & 255, ar(o)]; }
-function en_107(o) { return [107, o.v.length >>> 8, o.v.length & 255, ar(o)]; }
-function en_104(o) {
-  var l = o.v.length, r = [];
-  for (var i = 0; i < l; i++)r[i] = ein(o.v[i]);
-  return [104, l, r];
-}
-function unilen(o) {
-  return (o.v instanceof ArrayBuffer || o.v instanceof Uint8Array) ? o.v.byteLength :
-         (new TextEncoder().encode(o.v)).byteLength;
-}
-
-function en_109(o) {
-  var l = unilen(o);
-  return [109, l >>> 24, (l >>> 16) & 255, (l >>> 8) & 255, l & 255, ar(o)];
-}
-function en_108(o) {
-  var l = o.v.length, r = [];
-  for (var i = 0; i < l; i++)r.push(ein(o.v[i]));
-  return o.v.length == 0 ? [106] :
-         [108, l >>> 24, (l >>> 16) & 255, (l >>> 8) & 255, l & 255, r, 106];
-}
-function en_116(o) {
-  var l=o.v.length,x=[],r = [];
-  for (var i = 0; i < l; i++) r.push([ein(o.v[i].k),ein(o.v[i].v)]);
-  x = [116, l >>> 24, (l >>> 16) & 255, (l >>> 8) & 255, l & 255];
-  return o.v.length == 0 ? x : [x,r];
-}
-function en_110(o) {
-  var s=int_to_bytes(o.v); return [110,s.length,(o.v<0)?1:0].concat(s);
-}
-
-// DECODER
-
-function nop(b) { return []; };
-function big(b) {
-  var sk = b == 1 ? sx.getUint8(ix++) : sx.getInt32((a = ix, ix += 4, a));
-  var ret = 0, sig = sx.getUint8(ix++), count = sk;
-  while (count-- > 0) { ret = 256 * ret + sx.getUint8(ix + count); }
-  ix += sk; return ret * (sig == 0 ? 1 : -1);
-}
-function int(b) {
-  return b == 1 ? sx.getUint8(ix++) : sx.getInt32((a = ix, ix += 4, a));
-}
-function dec(d) {
-  sx = new DataView(d); ix = 0;
-  if (sx.getUint8(ix++) !== 131) throw ("BERT?");
-  return din();
-}
-function str(b) {
-  var dv, sz = (b==2?sx.getUint16(ix):(b==1?sx.getUint8(ix):sx.getUint32(ix)));
-  ix += b; var r = sx.buffer.slice(ix, ix += sz);
-  return utf8_arr(r);
-}
-function run(b) {
-  var sz = (b == 1 ? sx.getUint8(ix) : sx.getUint32(ix)), r = []; ix += b;
-  for (var i = 0; i < sz; i++) r.push(din());
-  return r;
-}
-function rut(b) {
-  var sz = (b == 1 ? sx.getUint8(ix) : sx.getUint32(ix)), r = []; ix += b;
-  for (var i = 0; i < sz; i++) r.push(din()); din();
-  return r;
-}
-function dic(b) {
-  var sz = sx.getUint32(ix), r = []; ix += 4;
-  for (var i = 0; i < sz; i++) r.push({k:din(),v:din()});
-  return r;
-}
-function iee(x) {
-  return read_Float(new Uint8Array(sx.buffer.slice(ix,ix+=8)),0,false,52,8);
-}
-function flo(x) {
-  return parseFloat(utf8_arr(sx.buffer.slice(ix, ix += 31)));
-}
-function arr(b) {
-  var dv, sz = sx.getUint16(ix); ix += b;
-  return new Uint8Array(sx.buffer.slice(ix, ix += sz));
-}
-function ref(cr) {
-  var adj = sx.getUint8(ix++), d; adj += sx.getUint8(ix++);
-  d = din(); ix += cr+adj*4; return d;
-}
-
-function din() {
-  var c = sx.getUint8(ix++), x;
-  switch (c) {
-    case  97: x = [int, 1]; break; case  98: x = [int, 4]; break;
-    case  99: x = [flo, 0]; break; case  70: x = [iee, 0]; break;
-    case 100: x = [str, 2]; break; case 104: x = [run, 1]; break;
-    case 107: x = [arr, 2]; break; case 108: x = [rut, 4]; break;
-    case 109: x = [str, 4]; break; case 110: x = [big, 1]; break;
-    case 111: x = [big, 4]; break; case 115: x = [str, 1]; break;
-    case 118: x = [str, 2]; break; case 119: x = [str, 1]; break;
-    case 105: x = [run, 4]; break; case 116: x = [dic, 4]; break;
-     case 90: x = [ref, 4]; break; case 114: x = [ref, 1]; break;
-    default:  x = [nop, 0];
-  } return { t: c, v: x[0](x[1]) };
-}
-
-// HELPERS
-
-function int_to_bytes(Int) {
-  if (Int % 1 !== 0) return [0];
-  var isNegative, OriginalInt, i, Rem, s = [];
-  isNegative = (Int < 0);
-  if (isNegative) { Int = - Int - 1; }
-  OriginalInt = Int;
-  var length = 0;
-  while (Int !== 0) { Rem = Int % 256;
-    if (isNegative) { Rem = 255 - Rem; }
-    s.push(Rem); Int = Math.floor(Int / 256); length++; }
-  if (Int > 0) { throw ("Argument out of range: " + OriginalInt); }
-  return s;
-}
-function uc(u1, u2) {
-  if (u1.byteLength == 0) return u2; if (u2.byteLength == 0) return u1;
-  var a = new Uint8Array(u1.byteLength + u2.byteLength);
-  a.set(u1, 0); a.set(u2, u1.byteLength); return a;
-}
-function ar(o) {
-  return o.v instanceof ArrayBuffer ? new Uint8Array(o.v) : o.v instanceof Uint8Array ? o.v :
-    Array.isArray(o.v) ? new Uint8Array(o.v) : new Uint8Array(utf8_enc(o.v));
-}
-function fl(a) {
-  return a.reduce(function (f, t) {
-    return uc(f, t instanceof Uint8Array ? t :
-      Array.isArray(t) ? fl(t) : new Uint8Array([t]));
-  }, new Uint8Array());
-}
+(function(exports){"use strict";let sx,ix,a;function tuple(){return{t:104,v:Array.apply(null,arguments)}}function list(){return{t:108,v:Array.apply(null,arguments)}}function map(){return{t:116,v:Array.apply(null,arguments)}}function atom(n){return{t:100,v:utf8_enc(n)}}function string(n){return{t:107,v:utf8_enc(n)}}function float(n){return{t:70,v:n}}function number(n){const e=n%1==0;return e&&n>=0&&n<256?{t:97,v:n}:e&&n>=-134217728&&n<=134217727?{t:98,v:n}:{t:110,v:n}}function bin(n){return{t:109,v:n instanceof ArrayBuffer?new Uint8Array(n):n instanceof Uint8Array?n:utf8_enc(n)}}function enc(n){return fl([131,ein(n)])}function ein(o){return Array.isArray(o)?en_108({t:108,v:o}):eval("en_"+o.t)(o)}function en_undefined(n){return[106]}function en_70(n){var e=Array(8).fill(0).flat();return write_Float(e,n.v,0,!1,52,8),[70].concat(e)}function en_97(n){return[97,n.v]}function en_98(n){return[98,n.v>>>24,n.v>>>16&255,n.v>>>8&255,255&n.v]}function en_99(n){var e=n.v.toExponential(20),t=/([^e]+)(e[+-])(\d+)/.exec(e),r=1==t[3].length?"0"+t[3]:t[3],i=Array.from(bin(t[1]+t[2]+r).v);return[n.t].concat(i).concat(Array(31-i.length).fill(0).flat())}function en_106(n){return[106]}function en_115(n){return[115,n.v.length,ar(n)]}function en_119(n){return[119,ar(n).length,ar(n)]}function en_118(n){return[118,ar(n).length>>>8,255&ar(n).length,ar(n)]}function en_100(n){return[100,n.v.length>>>8,255&n.v.length,ar(n)]}function en_107(n){return[107,n.v.length>>>8,255&n.v.length,ar(n)]}function en_104(n){for(var e=n.v.length,t=[],r=0;r<e;r++)t[r]=ein(n.v[r]);return[104,e,t]}function unilen(n){return n.v instanceof ArrayBuffer||n.v instanceof Uint8Array?n.v.byteLength:(new TextEncoder).encode(n.v).byteLength}function en_109(n){var e=unilen(n);return[109,e>>>24,e>>>16&255,e>>>8&255,255&e,ar(n)]}function en_108(n){for(var e=n.v.length,t=[],r=0;r<e;r++)t.push(ein(n.v[r]));return 0===n.v.length?[106]:[108,e>>>24,e>>>16&255,e>>>8&255,255&e,t,106]}function en_116(n){for(var e,t=n.v.length,r=[],i=0;i<t;i++)r.push([ein(n.v[i].k),ein(n.v[i].v)]);return e=[116,t>>>24,t>>>16&255,t>>>8&255,255&t],0==n.v.length?e:[e,r]}function en_110(n){var e=int_to_bytes(n.v);return[110,e.length,n.v<0?1:0].concat(e)}function nop(n){return[]}function big(n){for(var e=1==n?sx.getUint8(ix++):sx.getInt32((a=ix,ix+=4,a)),t=0,r=sx.getUint8(ix++),i=e;i-- >0;)t=256*t+sx.getUint8(ix+i);return ix+=e,t*(0==r?1:-1)}function int(n){return 1==n?sx.getUint8(ix++):sx.getInt32((a=ix,ix+=4,a))}function dec(n){if(sx=new DataView(n),ix=0,131!==sx.getUint8(ix++))throw"BERT?";return din()}function str(n){var e=2===n?sx.getUint16(ix):1===n?sx.getUint8(ix):sx.getUint32(ix);ix+=n;var t=sx.buffer.slice(ix,ix+=e);return utf8_arr(t)}function run(n){var e=1===n?sx.getUint8(ix):sx.getUint32(ix),t=[];ix+=n;for(var r=0;r<e;r++)t.push(din());return t}function rut(n){var e=1===n?sx.getUint8(ix):sx.getUint32(ix),t=[];ix+=n;for(var r=0;r<e;r++)t.push(din());return din(),t}function dic(n){var e=sx.getUint32(ix),t=[];ix+=4;for(var r=0;r<e;r++)t.push({k:din(),v:din()});return t}function iee(n){return read_Float(new Uint8Array(sx.buffer.slice(ix,ix+=8)),0,!1,52,8)}function flo(n){return parseFloat(utf8_arr(sx.buffer.slice(ix,ix+=31)))}function arr(n){var e=sx.getUint16(ix);return ix+=n,new Uint8Array(sx.buffer.slice(ix,ix+=e))}function ref(n){var e,t=sx.getUint8(ix++);return t+=sx.getUint8(ix++),e=din(),ix+=n+4*t,e}function din(){let n,e=sx.getUint8(ix++);switch(e){case 97:n=[int,1];break;case 98:n=[int,4];break;case 99:n=[flo,0];break;case 70:n=[iee,0];break;case 100:n=[str,2];break;case 104:n=[run,1];break;case 107:n=[arr,2];break;case 108:n=[rut,4];break;case 109:n=[str,4];break;case 110:n=[big,1];break;case 111:n=[big,4];break;case 115:n=[str,1];break;case 118:n=[str,2];break;case 119:n=[str,1];break;case 105:n=[run,4];break;case 116:n=[dic,4];break;case 90:n=[ref,4];break;case 114:n=[ref,1];break;default:n=[nop,0]}return{t:e,v:n[0](n[1])}}function int_to_bytes(n){if(n%1!=0)return[0];var e,t,r,i=[];for((e=n<0)&&(n=-n-1),t=n;0!==n;)r=n%256,e&&(r=255-r),i.push(r),n=Math.floor(n/256);if(n>0)throw"Argument out of range: "+t;return i}function uc(n,e){if(0===n.byteLength)return e;if(0===e.byteLength)return n;var t=new Uint8Array(n.byteLength+e.byteLength);return t.set(n,0),t.set(e,n.byteLength),t}function ar(n){return n.v instanceof ArrayBuffer?new Uint8Array(n.v):n.v instanceof Uint8Array?n.v:Array.isArray(n.v)?new Uint8Array(n.v):new Uint8Array(utf8_enc(n.v))}function fl(n){return n.reduce((function(n,e){return uc(n,e instanceof Uint8Array?e:Array.isArray(e)?fl(e):new Uint8Array([e]))}),new Uint8Array)}exports.arr=arr,exports.atom=atom,exports.big=big,exports.bin=bin,exports.dec=dec,exports.dic=dic,exports.din=din,exports.ein=ein,exports.en_100=en_100,exports.en_104=en_104,exports.en_106=en_106,exports.en_107=en_107,exports.en_108=en_108,exports.en_109=en_109,exports.en_110=en_110,exports.en_115=en_115,exports.en_116=en_116,exports.en_118=en_118,exports.en_119=en_119,exports.en_70=en_70,exports.en_97=en_97,exports.en_98=en_98,exports.en_99=en_99,exports.en_undefined=en_undefined,exports.enc=enc,exports.flo=flo,exports.float=float,exports.iee=iee,exports.int=int,exports.list=list,exports.map=map,exports.nop=nop,exports.number=number,exports.ref=ref,exports.run=run,exports.rut=rut,exports.str=str,exports.string=string,exports.tuple=tuple,exports.unilen=unilen,Object.defineProperty(exports,"__esModule",{value:!0})})(this.window=this.window||{});
+//# sourceMappingURL=bert.js.map
